Add vagrant support
The idea here is that a user reluctant to install software on their
laptop directly could instead use Vagrant to create a VM containing
the software needed.

I kept the Vagrantfile as simple as possible and am re-using the
linux.sh script ( modified a bit, but should still have same
functionality if run without vagrant ) to install the packages.

Goal would be even to create/upload the Vagrant box to Vagrant Cloud
so that it could be used in other journeys/repos without having to copy
the install instructions into each of them.